requirements.txt
aiohttp==3.8.1
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
typing-extensions==4.2.0
yarl==1.7.2
numpy==1.23.0
num2words~=0.5.10